Create common agents list Vue component
